refactor(miden, processor): changing type of clk to u32 to optimise memory

+++ miden/src/tools/mod.rs
@@ -45,16 +45,15 @@ impl Analyze {
 ///   the number of vm cycles it takes to execute the instruction and the number of times the
 ///   instruction is run as part of the given program.
 #[derive(Debug, Default, Eq, PartialEq)]
-// TODO: total_vm_cycles cannot exceed clk and can therefore also be casted as u32?
 pub struct ProgramInfo {
-    total_vm_cycles: usize,
+    total_vm_cycles: u32,
     total_noops: usize,
     asm_op_stats: Vec<AsmOpStats>,
 }
 
 impl ProgramInfo {
     /// Returns total vm cycles to execute a program
-    pub fn total_vm_cycles(&self) -> usize {
+    pub fn total_vm_cycles(&self) -> u32 {
         self.total_vm_cycles
     }
 
@@ -79,7 +78,7 @@ impl ProgramInfo {
 
     /// Sets the total vm cycles to the provided value
     pub fn set_total_vm_cycles(&mut self, total_vm_cycles: u32) {
-        self.total_vm_cycles = total_vm_cycles as usize;
+        self.total_vm_cycles = total_vm_cycles;
     }
 
     /// Records a new occurrence of asmop in the sorted asmop stats vector of this program info.
